Extract plot_haplotype_network into AnophelesHapNetAnalysis mixin (#366)
